refactor: Config 로딩 로직 유틸리티 모듈화
